+ bugfix: prawdziwy bugfix NaNów w kamerze + bugfix: zgranie oświetlenia atmosfery
git-svn-id: svn+ssh://gamma/home/samba/koturj/svn/inz@281 e70bc5d4-e669-4a26-a90f-bfde2a246bde

+++ src/ui/camera.cpp
@@ -32,7 +32,7 @@ void CamFreeLook::born( Matrix state , void*data )
 void CamFreeLook::learn( void * data )
 {
 	if( !data ) return;
-	move_speed = *(float*)data;
+	move_speed = *(double*)data;
 }
 
 UI::Camera::Matrix CamFreeLook::work()
